Polish VPI_Vmod_Init() signature
Found using vclflint

Upstream: https://code.vinyl-cache.org/vinyl-cache/vinyl-cache/commit/231cc3305a1a945337ba0668eaad345b218fb2ee
Port: exact

+++ include/vcc_interface.h
@@ -72,7 +72,7 @@ struct wrk_vpi {
 void VPI_trace(VRT_CTX, unsigned);
 void VPI_vcl_fini(VRT_CTX);
 
-int VPI_Vmod_Init(VRT_CTX, struct vmod **hdl, unsigned nbr, void *ptr, int len,
+int VPI_Vmod_Init(VRT_CTX, struct vmod **hdl, unsigned nbr, void *ptr, size_t len,
     const char *nm, const char *path, const char *file_id, const char *backup);
 void VPI_Vmod_Unload(VRT_CTX, struct vmod **hdl);
